% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tarifas.R
\name{data_arquivo_tarifas}
\alias{data_arquivo_tarifas}
\title{Obter a data do arquivo oficial de tarifas}
\usage{
data_arquivo_tarifas()
}
\value{
Um objeto de classe \code{Date} com a data encontrada no link do arquivo oficial.
}
\description{
Consulta o link do arquivo oficial de tarifas vigentes do Brasil no site da
Camex e retorna a data presente no nome do arquivo.
}
\examples{
\dontrun{
data_arquivo_tarifas()
}

}
