% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detalhar_listas_excecao_vigentes.R
\name{detalhar_listas_excecao_vigentes}
\alias{detalhar_listas_excecao_vigentes}
\title{Consolida os codigos NCM em listas de excecao tarifaria e adiciona colunas de
contagem de quota e de destaque (Ex) bem como uma coluna para indicar a presença
de NCM integral.}
\usage{
detalhar_listas_excecao_vigentes(x)
}
\arguments{
\item{x}{deve ser o resultado da funcao \code{download_tarifas()}.}
}
\value{
um \code{tibble} (ou \code{data.frame}) com uma linha por NCM e as colunas:
\itemize{
\item \code{ncm}: codigo NCM;
\item \code{contagem_quota}: numero de ocorrencias com quota associadas a NCM nas
listas de excecao vigentes;
\item \code{contagem_ex}: numero de ocorrencias com destaque (Ex) associadas a NCM
nas listas de excecao vigentes;
\item \code{ncm_integral}: indicador binario (0/1) informando se existe medida em lista
de exceção que abrange a NCM integralmente (sem destaque);
\item \code{lista}: string com a(s) lista(s) de excecao em que a NCM aparece,
concatenadas com virgula quando houver mais de uma.
\item \code{tarifa_aplicada}: apresenta a tarifa aplicada para as ncms incluidas
integralmente em lista de excecao.
}
}
\description{
Itera sobre as seguintes listas de excecao tarifaria: desabastecimento
(\strong{IV}), Letec (\strong{V}), Lebitbk (\strong{VI}), Concessoes OMC
(\strong{VIII}), DCC (\strong{IX}) e Automotivos ACE-14 (\strong{X}) e produz
um resumo contendo todos os codigos NCM vigentes nessas listas com contagem
de quota e de destaques (Ex), indicador de NCM integral, lista(s) de excecao
e tarifa aplicada quando uma NCM for incluída integralmente em uma lista sem
que seja uma quota.
}
\details{
A funcao aplica, para cada anexo especificado, a combinacao de
\code{adiciona_indicador_ex_quota} e \code{tarifas_vigentes} para:
\itemize{
\item ler e formatar o anexo correspondente;
\item adicionar indicadores de quota, destaque (Ex) e NCM integral; e
\item filtrar apenas os registros vigentes na data de execucao (\code{Sys.Date()}).
}

Os anexos tambem sao processados para obter as ncms incluidas integralmente
em listas de excecao com suas respectivas tarifas aplicadas.

Em seguida, os resultados de cada anexo sao agregados por NCM e consolidados
em um unico \code{tibble}/\code{data.frame}. Por fim, sao adicionados dados de tarifas
aplicadas a partir das listas de excecao, produzindo:
\itemize{
\item o numero de ocorrencias com quota por NCM na lista de excecao especificada;
\item o numero de destaques (Ex) por NCM na lista de excecao especificada;
\item o indicador binario (0/1) informando se ha medida que abrange a NCM
integralmente (sem destaque);
\item a coluna \code{lista}, que mantem a(s) lista(s) de excecao em que a NCM aparece;
\item a coluna \code{tarifa_aplicada} com dados apenas para as ncms incluidas
integralmente em listas de excecao. Pode ocorrer de uma NCM estar presente
em mais de uma lista de excecao, mas, nesse caso, em regra, a NCM aparece
na lista de desabastecimento e em outra lista ou na lista de concessoes
da OMC e em outra lista. Como normalmente os codigos presentes na lista de
desabastecimento e de concessoes da OMC nao sao incluidos nessas
listas integralmente, nesses casos de uma NCM fazer parte de mais de uma lista,
a tarifa aplicada sera, provavelmente, referente a outra lista remanescente.
}
}
\examples{
# Exemplo de uso:
x <- download_tarifas()
resumo_excecoes <- detalhar_listas_excecao_vigentes(x)
dplyr::glimpse(resumo_excecoes)

}
