% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tarifas.R
\name{download_tarifas}
\alias{download_tarifas}
\title{Download de tarifas de importação do Brasil do site oficial da
Camara de Comercio Exterior (Camex)}
\usage{
download_tarifas(destfile = NULL)
}
\arguments{
\item{destfile}{Caminho do arquivo onde o download deve ser salvo. Deve
incluir o nome do arquivo, e o diretório informado deve existir. Quando
\code{NULL}, usa um arquivo temporário. Se o arquivo já existir, ele será
sobrescrito.}
}
\value{
O caminho onde o arquivo oficial de tarifas foi salvo.
}
\description{
Essa função faz o download do arquivo de tarifas vigentes do Brasil. Por
padrão, o arquivo é salvo em um diretório temporário. Caso \code{destfile} seja
informado, o arquivo é salvo no caminho indicado.
}
