% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tarifas.R
\name{extrair_data_link_tarifas}
\alias{extrair_data_link_tarifas}
\title{Extrair a data do link do arquivo oficial de tarifas}
\usage{
extrair_data_link_tarifas(link)
}
\arguments{
\item{link}{String com o link do arquivo oficial de tarifas.}
}
\value{
Um objeto de classe \code{Date} com a data encontrada no link.
}
\description{
Funcao auxiliar para extrair a data no formato \code{dd-mm-aaaa} presente no link
do arquivo oficial de tarifas.
}
\keyword{internal}
