% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_anexo1.R
\name{ler_anexo1}
\alias{ler_anexo1}
\title{Função auxiliar usada na função ler_anexo}
\usage{
ler_anexo1(x)
}
\arguments{
\item{x}{Objeto retornado pela função \code{download_tarifas}}
}
\value{
Um tibble que retorna as colunas ncm, descricao_tec,
descricao_tec_concatenada, tec, bkbit, resolucoes.
}
\description{
Esta função faz a leitura e processamento do Anexo I para obtenção
da TEC a partir do arquivo oficial de tarifas disponibilizado no site da Camex.
}
\details{
Lê a aba referente ao Anexo I da planiha excel obtida pela função
\code{download_tarifas}, limpa e organiza as informações.

A função cria a coluna descricao_tec_concatenada, a qual apresenta a concatenação
das descrições que vão da posição (4 dígitos) até o subitem da NCM (8 dígitos).
A coluna tec representa a Tarifa Externa Comum e os seus valores estão em
porcentagens.
}
\keyword{internal}
