% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_anexo_formatado.R
\name{listar_anexos}
\alias{listar_anexos}
\title{Função para consultar nome e número dos anexos das tarifas de importação}
\usage{
listar_anexos()
}
\value{
Retorna um tibble com número do anexo e nome correspondente do anexo
}
\description{
Função pode ser utilizada para consulta rápida dos nomes e números dos anexos
de tarifas ou para verificar qual argumento usar na função \code{ler_anexo_formatado}.
}
\examples{
listar_anexos()

}
