% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tarifas.R
\name{obter_link_arquivo_tarifas}
\alias{obter_link_arquivo_tarifas}
\title{Obter o link do arquivo oficial de tarifas}
\usage{
obter_link_arquivo_tarifas()
}
\value{
Uma string com o link do arquivo oficial de tarifas.
}
\description{
Funcao auxiliar para consultar a pagina oficial de tarifas vigentes e
identificar o link do arquivo Excel com os anexos tarifarios.
}
\keyword{internal}
