% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tarifas_vigentes.R
\name{tarifas_futuras}
\alias{tarifas_futuras}
\title{Filtra as tarifas futuras de listas de excecao}
\usage{
tarifas_futuras(x)
}
\arguments{
\item{x}{uma lista de excecao obtida a partir da funcao \code{ler_anexo}.
A funcao utiliza a coluna \code{inicio_de_vigencia} no formato \code{Date}
para obter as tarifas futuras.}
}
\value{
o mesmo objeto de entrada, porem filtrado para conter apenas
as tarifas futuras na data de execucao da funcao.
}
\description{
Retorna apenas os registros cuja tarifa entrara em vigencia no futuro. Compara
a data da execucao da funcao com a data na coluna \code{inicio_de_vigencia} para
filtrar as tarifas futuras presentes em listas de excecao.
No caso de nao haver medidas em datas futuras, a funcao retornara um tibble
vazio.
}
\examples{
\dontrun{
x <- download_tarifas()
desabastecimento <- ler_anexo(x, "iv") # ler anexoi iv - desabastecimento
tarifas_futuras(desabastecimento)
}

}
