% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tarifas_vigentes.R
\name{tarifas_vigentes}
\alias{tarifas_vigentes}
\title{Filtra as tarifas vigentes de listas de excecao}
\usage{
tarifas_vigentes(x)
}
\arguments{
\item{x}{uma lista de excecao obtida a partir da funcao \code{ler_anexo}.
A funcao utiliza as colunas \code{inicio_de_vigencia} e \code{termino_de_vigencia}
no formato \code{Date} para obter as tarifas vigentes.}
}
\value{
O mesmo objeto de entrada, porem filtrado para conter apenas
as tarifas vigentes na data de execucao da funcao.
}
\description{
Retorna apenas os registros cuja tarifa esta \strong{vigente hoje},
considerando o intervalo entre \code{inicio_de_vigencia} e \code{termino_de_vigencia}.
A depender da lista consultada e da data pode ou nao haver medidas em datas
futuras. No caso de nao haver medidas em datas futuras, nao havera diferenca
em aplicar ou nao esta funcao, uma vez que os dados serao identicos.
}
\examples{
\dontrun{
x <- download_tarifas()
desabastecimento <- ler_anexo(x, "iv") # ler anexoi iv - desabastecimento
tarifas_vigentes(desabastecimento)
}

}
