% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tarifasimpbr-package.R
\docType{package}
\name{tarifasimpbr-package}
\alias{tarifasimpbr}
\alias{tarifasimpbr-package}
\title{tarifasimpbr: Ferramentas para trabalhar com dados de tarifas de importação brasileiras}
\description{
Fornece funções para baixar o arquivo de tarifas oficial da SE-Camex, ler os anexos desse arquivo de forma padronizada e compilar os anexos para obter a tarifa aplicada vigente.
}
\author{
\strong{Maintainer}: Miguel Cleaver \email{miguelmdic@outlook.com}

}
\keyword{internal}
