% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tarifas.R
\name{download_tarifas}
\alias{download_tarifas}
\title{Download de tarifas de importação do Brasil do site oficial da
Camara de Comerio Exterior (Camex)}
\usage{
download_tarifas()
}
\value{
o caminho temporário, onde o arquivo oficial de tarifas foi salvo
}
\description{
Essa função faz o download do arquivo de tarifas vigentes do Brasil em um diretorio
temporario.
}
