% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_anexo.R
\name{listar_anexos}
\alias{listar_anexos}
\title{Funcao para consultar nome e numero dos anexos das tarifas de importacao}
\usage{
listar_anexos()
}
\value{
Retorna um tibble com numero do anexo e nome correspondente do anexo
}
\description{
Funcao pode ser utilizada para consulta rapida dos nomes e numeros dos anexos
de tarifas ou para verificar qual argumento usar na funcao \code{ler_anexo}.
}
\examples{
listar_anexos()

}
