% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tarifas_vigentes.R
\name{tarifas_futuras}
\alias{tarifas_futuras}
\title{Filtra as tarifas futuras de listas de exceção}
\usage{
tarifas_futuras(x)
}
\arguments{
\item{x}{uma lista de exceção obtida a partir da função \code{ler_anexo_formatado}.
A função utiliza as colunas \code{inicio_de_vigencia} e \code{termino_de_vigencia}
no formato \code{Date} para obter as tarifas vigentes.}
}
\value{
o mesmo objeto de entrada, porém filtrado para conter apenas
as tarifas vigentes na data de execução da função.
}
\description{
Retorna apenas os registros cuja tarifa entrará em vigência no futuro. Compara
a data da execução da função com a data na coluna \code{inicio_de_vigencia} para
filtrar as tarifas futuras presentes em listas de exceção.
No caso de não haver medidas em datas futuras, a função retornará um tibble
vazio.
}
\examples{
\dontrun{
x <- download_tarifas()
desabastecimento <- ler_anexo_formatado(x, "iv") # ler anexoi iv - desabastecimento
tarifas_futuras(desabastecimento)
}

}
