% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tarifas_vigentes.R
\name{tarifas_vigentes}
\alias{tarifas_vigentes}
\title{Filtra as tarifas vigentes de listas de exceção}
\usage{
tarifas_vigentes(x)
}
\arguments{
\item{x}{uma lista de exceção obtida a partir da função \code{ler_anexo_formatado}.
A função utiliza as colunas \code{inicio_de_vigencia} e \code{termino_de_vigencia}
no formato \code{Date} para obter as tarifas vigentes.}
}
\value{
O mesmo objeto de entrada, porém filtrado para conter apenas
as tarifas vigentes na data de execução da função.
}
\description{
Retorna apenas os registros cuja tarifa está \strong{vigente hoje},
considerando o intervalo entre \code{inicio_de_vigencia} e \code{termino_de_vigencia}.
A depender da lista consultada e da data pode ou não haver medidas em datas
futuras. No caso de não haver medidas em datas futuras, não haverá diferença
em aplicar ou não está função, uma vez que os dados serão idênticos.
}
\examples{
\dontrun{
x <- download_tarifas()
desabastecimento <- ler_anexo_formatado(x, "iv") # ler anexoi iv - desabastecimento
tarifas_vigentes(desabastecimento)
}

}
